speedmatrix: AsNumber reads sized integers so an in-memory metrics row passes the row check (#621)
The first rendered speedmatrix/plain run after 51257a32 failed its own
row validation: the case reads its metrics rows back before JSON ever
touched them, so ticks_advanced arrived as uint64 and AsNumber read it
as -1 (its missing-value result). AsNumber now converts every sized
integer and float32; the row test carries the integer form.

Rendered speedmatrix/plain passed with all eight rows
(.rimgovernor/out/speedmatrix-621b): governor-off 5247 wall TPS,
uncapped 3266, viewer 2221 at 7.7 frames/s; paused_ms 0 on every row
against a status-sample ratio of up to 0.15, since no row stopped
within its budget.

+++ go/internal/nativeaccept/speedmatrix_test.go
@@ -205,8 +205,10 @@ func TestSpeedRowProblems(t *testing.T) {
 	if got := CompareOutcomes(nil, 1); len(got) != 0 {
 		t.Fatalf("CompareOutcomes is not the boundary: %v", got)
 	}
+	// A row read back in memory carries its tick count as the integer the
+	// case computed, never the float64 JSON would decode.
 	full := SpeedMetricsFromRows([]map[string]any{
-		{"case": "Normal", "speed": "Normal", "wall_tps": 60.0, "ticks_advanced": 2300.0},
+		{"case": "Normal", "speed": "Normal", "wall_tps": 60.0, "ticks_advanced": uint64(2300)},
 		{"case": "uncapped", "speed": "Ultrafast", "wall_tps": 4000.0, "ticks_advanced": 30000.0},
 		{"case": "governor-off", "speed": "Ultrafast", "wall_tps": 5000.0, "ticks_advanced": 30000.0},
 		{"case": "viewer", "speed": "Ultrafast", "wall_tps": 3500.0, "ticks_advanced": 30000.0},
